normalize file names
